Resetting blog styling to how it was before (please read comment)
Also moved script to bottom of page for better loading. Bootstrap seems quite sensitive to playing with the layout so before I kept it as simple as possible. It seems the last changes broke the layout so I’ve reset it and added the changes I think you were trying to achieve (smaller, centred user image on previous posts).

PS. Col width should add to 12 (which is why I’ve reset those). Also I’ve reset the header to a row as this keeps the layout consistent